Add files via upload
Creating a program that requests the name of four football teams and simulates matches so that each team plays the other three once. During the match, you must ask how many goals each team scored and add up the appropriate scores. At the end, you must say which teams were champions, as there may be a tie in points. Note: victory is worth 3 points to the winner, a draw is worth 1 point to each team and defeat does not add points.